Renaming and comment
