+++ src/hotspot/share/runtime/deoptimization.cpp
@@ -61,10 +61,10 @@
 #include "prims/methodHandles.hpp"
 #include "prims/vectorSupport.hpp"
 #include "runtime/atomic.hpp"
+#include "runtime/clearFrameAnchorMark.hpp"
 #include "runtime/continuation.hpp"
 #include "runtime/continuationEntry.inline.hpp"
 #include "runtime/deoptimization.hpp"
-#include "runtime/disableStackTracingMark.hpp"
 #include "runtime/escapeBarrier.hpp"
 #include "runtime/fieldDescriptor.hpp"
 #include "runtime/fieldDescriptor.inline.hpp"
@@ -756,7 +756,7 @@ JRT_LEAF(BasicType, Deoptimization::unpack_frames(JavaThread* thread, int exec_m
     // We set the last_Java frame. But the stack isn't really parsable here. So we
     // clear it to make sure JFR understands not to try and walk stacks from events
     // in here.
-    DisableStackTracingMark dstm(thread);
+    ClearFrameAnchorMark dstm(thread);
 
     // Unpack the interpreter frames and any adapter frame (c2 only) we might create.
     array->unpack_to_stack(stub_frame, exec_mode, info->caller_actual_parameters());
